dense matrix float type for gaussian filter

+++ hicprediction/predict.py
@@ -239,7 +239,7 @@ def predictionToMatrix(pred, baseFilePath, pModelParams, chromosome, predictionF
         if pSigma > 0.0:
             upper = sparse.triu(predMatrix)
             lower = sparse.triu(predMatrix, k=1).T
-            fullPredMatrix = (upper+lower).todense().astype('uint32')
+            fullPredMatrix = (upper+lower).todense().astype('float32')
             filteredPredMatrix = ndimage.gaussian_filter(fullPredMatrix,pSigma)
             predMatrix = sparse.triu(filteredPredMatrix)
